refactor(DEVX-7053): rename ghes:register to vcs:github-host:add
Move command from Commands/Ghes/RegisterCommand to
Commands/Vcs/GithubHost/AddCommand to follow existing VCS command
namespace conventions.
